Presenter Refactor
