+++ AppFramework/build.gradle
@@ -7,8 +7,8 @@ android {
     defaultConfig {
         minSdkVersion 16
         targetSdkVersion 23
-        versionCode 1
-        versionName "1.0.1"
+        versionCode 2
+        versionName "1.1.0"
     }
     buildTypes {
         release {
