--- AppFramework/src/main/res/layout-land/nav_bar.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:id="@+id/nav_bar_layout"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="match_parent"
-    android:layout_gravity="left"
-    style="@android:style/ButtonBar"
-    android:orientation="vertical">
-
-</LinearLayout>
