--- AppFramework/src/main/res/layout/nav_bar.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:id="@+id/nav_bar_layout"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="bottom"
-    style="@android:style/ButtonBar"
-    android:orientation="horizontal">
-
-</LinearLayout>
