--- AppFramework/src/main/res/layout/nav_button.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:id="@+id/nav_button_layout"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="bottom"
-    style="@android:style/ButtonBar"
-    android:paddingBottom="5dp"
-    android:background="@android:color/transparent"
-    android:orientation="vertical">
-
-    <ImageView
-        android:id="@+id/nav_button_image"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:scaleType="fitCenter"/>
-
-    <TextView
-        android:id="@+id/nav_button_text"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"/>
-
-</LinearLayout>
