--- AppFramework/src/main/res/values/ids.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <item type="id" name="pager_view" />
-</resources>
