+++ AppFramework/src/main/res/values/strings.xml
@@ -13,6 +13,4 @@
     <string name="edit_view_ok_button_new">Create</string>
     <string name="edit_view_button_cancel">Cancel</string>
 
-    <string name="menu_refresh">Refresh</string>
-
 </resources>
