+++ MockTradeApp/src/main/AndroidManifest.xml
@@ -21,9 +21,18 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <activity
-            android:name="com.balch.android.app.framework.domain.EditActivity"
-            android:label="You Should not see this" >
+            android:name="com.balch.android.app.framework.domain.EditActivity" >
+        </activity>
+
+        <activity
+            android:name=".order.OrderListActivity"
+            android:label="@string/order_list_activity_label" >
+            android:parentActivityName=".MainActivity" >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".MainActivity" />
         </activity>
 
         <receiver
