--- MockTradeApp/src/main/res/layout/portfolio_view.xml
@@ -1,13 +1,0 @@
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/portfolio_view_layout"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    <ExpandableListView
-        android:id="@+id/portfolio_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:groupIndicator="@null"/>
-</LinearLayout>
